Move single replica plot to separate file
